refactor: add entityField type
