+++ packages/visual-editor/src/components/contentBlocks/BodyText.tsx
@@ -46,12 +46,13 @@ const bodyTextFields: YextFields<BodyTextProps> = {
   data: YextField(msg("fields.data", "Data"), {
     type: "object",
     objectFields: {
-      text: YextField(msg("fields.text", "Text"), {
+      text: {
         type: "entityField",
+        label: msg("fields.text", "Text"),
         filter: {
           types: ["type.rich_text_v2"],
         },
-      }),
+      },
     },
   }),
   styles: YextField(msg("fields.styles", "Styles"), {
