+++ packages/visual-editor/src/components/contentBlocks/image/Image.tsx
@@ -63,15 +63,13 @@ export const ImageWrapperFields: YextFields<ImageWrapperProps> = {
   data: YextField(msg("fields.data", "Data"), {
     type: "object",
     objectFields: {
-      image: YextField<
-        any,
-        ImageType | ComplexImageType | TranslatableAssetImage
-      >(msg("fields.options.image", "Image"), {
+      image: {
         type: "entityField",
+        label: msg("fields.options.image", "Image"),
         filter: {
           types: ["type.image"],
         },
-      }),
+      },
       link: {
         type: "translatableString",
         label: msg("fields.link", "Link"),
