+++ packages/visual-editor/src/editor/README.md
@@ -86,30 +86,33 @@ import { EntityField } from "@yext/visual-editor";
 </EntityField>;
 ```
 
-## YextEntityFieldSelector
+## entityField Field Type
 
-Use this to allow Visual Editor users to choose an entity field or constant value that will populate data into a component.
+Use the registered `entityField` field type to allow Visual Editor users to choose an entity field or a constant value that will populate data into a component.
 The user can choose an entity field from a dropdown or use a constant value. Regardless, the user should always
 enter a constant value as it will be used as a fallback value in the case that the entity is missing the selected entity field.
+`YextEntityFieldSelector(...)` remains available as a compatibility wrapper, but new configs should author `entityField` directly in `YextFields`.
 
 The constant value field currently has limited functionality with complex object entity types. When using complex
 object types, ensure your render function handles undefined fields.
 
 ### Props
 
-| Name                 | Type            | Description                                                    |
-| -------------------- | --------------- | -------------------------------------------------------------- |
-| label?               | string          | The user-facing label for the field.                           |
-| filter.types         | string[]        | Determines which fields will be available based on field type. |
-| filter.allowList?    | types: string[] | Field names to include. Cannot be combined with disallowList.  |
-| filter.disallowList? | types: string[] | Field names to exclude. Cannot be combined with allowList.     |
+| Name                       | Type                | Description                                                    |
+| -------------------------- | ------------------- | -------------------------------------------------------------- |
+| type                       | `"entityField"`     | Registers the field as the entity/constant selector.           |
+| label?                     | string \| MsgString | The user-facing label for the field.                           |
+| filter.types               | string[]            | Determines which fields will be available based on field type. |
+| filter.allowList?          | string[]            | Field names to include. Cannot be combined with disallowList.  |
+| filter.disallowList?       | string[]            | Field names to exclude. Cannot be combined with allowList.     |
+| disableConstantValueToggle | boolean             | Disables static-content mode when true.                        |
+| disallowTranslation        | boolean             | Uses non-localized constant editors when supported.            |
 
 ### Usage
 
 ```tsx
 import {
   EntityFieldType,
-  YextEntityFieldSelector,
   resolveYextEntityField,
   useDocument,
   YextComponentConfig
@@ -129,14 +132,15 @@ const exampleFields: YextFields<ExampleProps> = {
     type: "object",
     label: "Example Parent Field", // top-level sidebar label
     objectFields: {
-      entityField: YextEntityFieldSelector<typeof config>({
+      entityField: {
+        type: "entityField",
         label: "Example Field", // sidebar label for the sub field
         filter: {
           types: ["type.string"],
           disallowList: ["exampleField"],
           //allowList: ["exampleField"],
         },
-      }),
+      },
     },
   },
 };
@@ -268,13 +272,12 @@ const myComponentFields: YextFields<MyComponentProps> = {
 
 ## YextField
 
-`YextField` provides a unified utility for creating typed field configurations in a [Puck](https://github.com/measuredco/puck) and Yext Visual Editor integration context. It abstracts over a subset of common field types and includes special handling for the internal `basicSelector` field type and [YextEntityFieldSelector](##YextEntityFieldSelector).
+`YextField` provides a unified utility for creating typed field configurations in a [Puck](https://github.com/measuredco/puck) and Yext Visual Editor integration context. It abstracts over a subset of common field types and includes special handling for internal helper field types such as `dynamicSelect`.
 
 ### Features
 
 - Strongly typed helper for defining field configs
 - Support for standard Puck field types such as `text`, `array`, and `object`
-- Extended support for Yext-specific entity selectors
 - Native Puck fields like `radio` can be authored directly inside `YextFields`
 
 ### Props
@@ -301,10 +304,11 @@ import {
 } from "@yext/visual-editor";
 
 const myComponentFields: YextFields<myComponentProps> = {
-  address: YextField<any, AddressType>(msg("fields.address", "Address"), {
+  address: {
     type: "entityField",
+    label: msg("fields.address", "Address"),
     filter: { types: ["type.address"] },
-  }),
+  },
   showGetDirections: {
     label: msg("fields.showGetDirections", "Show Get Directions Link"),
     type: "radio",
@@ -463,19 +467,21 @@ card: YextField(msg("fields.card", "Card"), {
 
 #### Entity Selector Field
 
-Renders a Yext entity selector with filtering capabilities.
+Use the registered `entityField` type to render the entity selector with filtering capabilities.
 
 ```tsx
-linkedEntity: YextField("Linked Entity", {
+linkedEntity: {
   type: "entityField",
-  filter: { entityType: "faq" },
-});
+  label: msg("fields.linkedEntity", "Linked Entity"),
+  filter: { types: ["type.string"] },
+},
 ```
 
 **Props:**
 
 - `type`: `"entityField"`
-- `filter`: `any` — passed to [YextEntityFieldSelector](##YextEntityFieldSelector)
+- `label?`: `string | MsgString`
+- `filter`: `RenderEntityFieldFilter`
 
 ---
