+++ packages/visual-editor/src/editor/YextField.test.tsx
@@ -4,24 +4,13 @@ import { YextAutoField } from "../fields/YextAutoField.tsx";
 import { msg } from "../utils/i18n/platform.ts";
 import { afterEach, describe, expect, it, vi } from "vitest";
 import { YextField } from "./YextField.tsx";
-const { dynamicOptionsSelectorMock, yextEntityFieldSelectorMock } = vi.hoisted(
-  () => ({
-    dynamicOptionsSelectorMock: vi.fn(),
-    yextEntityFieldSelectorMock: vi.fn(),
-  })
-);
+const { dynamicOptionsSelectorMock } = vi.hoisted(() => ({
+  dynamicOptionsSelectorMock: vi.fn(),
+}));
 
 vi.mock("./DynamicOptionsSelector.tsx", () => ({
   DynamicOptionsSelector: dynamicOptionsSelectorMock,
 }));
-
-vi.mock("./YextEntityFieldSelector.tsx", async () => {
-  const actual = await vi.importActual("./YextEntityFieldSelector.tsx");
-  return {
-    ...actual,
-    YextEntityFieldSelector: yextEntityFieldSelectorMock,
-  };
-});
 
 const renderCustomField = (field: any, value?: any) => {
   const onChange = vi.fn();
@@ -94,29 +83,6 @@ describe("YextField", () => {
 
     expect(document.querySelector("textarea")).toBeDefined();
     expect(screen.getByDisplayValue("Beta")).toBeDefined();
-  });
-
-  it("delegates entityField configs to YextEntityFieldSelector", () => {
-    const returnedField = createCustomField();
-    const fieldName = msg("fields.entity", "Entity Field");
-    const filter = { types: ["type.string"] };
-
-    yextEntityFieldSelectorMock.mockReturnValue(returnedField);
-
-    const field = YextField(fieldName, {
-      type: "entityField",
-      filter,
-      disableConstantValueToggle: true,
-      disallowTranslation: true,
-    } as any);
-
-    expect(yextEntityFieldSelectorMock).toHaveBeenCalledWith({
-      label: fieldName,
-      filter,
-      disableConstantValueToggle: true,
-      disallowTranslation: true,
-    });
-    expect(field).toBe(returnedField);
   });
 
   it("passes registered custom field configs through with the provided label", () => {
