+++ packages/visual-editor/src/internal/puck/components/meta-title/MetaTitleField.tsx
@@ -1,10 +1,11 @@
 import React from "react";
-import { AutoField, type Field } from "@puckeditor/core";
+import { type Field } from "@puckeditor/core";
 import { useDocument } from "../../../../hooks/useDocument.tsx";
 import { msg, pt } from "../../../../utils/i18n/platform.ts";
 import { getPageSetLocales } from "../../../../utils/pageSetLocales.ts";
 import { type YextEntityField } from "../../../../editor/YextEntityFieldSelector.tsx";
-import { YextField } from "../../../../editor/YextField.tsx";
+import { YextAutoField } from "../../../../fields/YextAutoField.tsx";
+import { type EntityFieldTypes } from "../../../../internal/utils/getFilteredEntityFields.ts";
 import { type TranslatableString } from "../../../../types/types.ts";
 import { getMetaTitleMissingLocales } from "./metaTitleValidation.ts";
 
@@ -25,19 +26,23 @@ export const MetaTitleField = (): Field<
       );
 
       const metaTitleField = React.useMemo(
-        () =>
-          YextField(msg("fields.metaTitle", "Meta Title"), {
-            type: "entityField",
-            filter: {
-              types: ["type.string"],
-            },
-          }),
+        () => ({
+          type: "entityField" as const,
+          label: msg("fields.metaTitle", "Meta Title"),
+          filter: {
+            types: ["type.string"] as EntityFieldTypes[],
+          },
+        }),
         []
       );
 
       return (
         <div className="ve-w-full">
-          <AutoField field={metaTitleField} value={value} onChange={onChange} />
+          <YextAutoField
+            field={metaTitleField}
+            value={value}
+            onChange={onChange}
+          />
           {missingLocales.length > 0 && (
             <p className="ve-mt-2 ve-text-xs ve-text-red-600" role="alert">
               {errorMessage}
